Move app native calls into OsApp
