+++ realm/realm-library/src/objectServer/java/io/realm/internal/objectstore/OsApp.java
@@ -1,15 +1,23 @@
 package io.realm.internal.objectstore;
 
+import java.util.Map;
+import java.util.concurrent.atomic.AtomicReference;
+
+import javax.annotation.Nullable;
+
+import io.realm.internal.KeepMember;
 import io.realm.internal.NativeObject;
+import io.realm.internal.jni.OsJNIResultCallback;
+import io.realm.internal.network.OkHttpNetworkTransport;
+import io.realm.internal.network.ResultHandler;
+import io.realm.mongodb.AppConfiguration;
+import io.realm.mongodb.AppException;
 
 public class OsApp implements NativeObject {
     private static final long nativeFinalizerPtr = nativeGetFinalizerMethodPtr();
 
+    private OsJavaNetworkTransport networkTransport;
     private final long nativePtr;
-
-    public OsApp(long nativePtr) {
-        this.nativePtr = nativePtr;
-    }
 
     @Override
     public long getNativePtr() {
@@ -19,6 +27,32 @@ public class OsApp implements NativeObject {
     @Override
     public long getNativeFinalizerPtr() {
         return nativeFinalizerPtr;
+    }
+
+    public OsApp(AppConfiguration config, String userAgentBindingInfo, String appDefinedUserAgent, String syncDir) {
+        nativePtr = nativeCreate(
+                config.getAppId(),
+                config.getBaseUrl().toString(),
+                config.getAppName(),
+                config.getAppVersion(),
+                config.getRequestTimeoutMs(),
+                config.getEncryptionKey(),
+                syncDir,
+                userAgentBindingInfo,
+                appDefinedUserAgent,
+                "android",
+                android.os.Build.VERSION.RELEASE,
+                io.realm.BuildConfig.VERSION_NAME);
+
+        this.networkTransport = new OkHttpNetworkTransport(config.getHttpLogObfuscator());
+        networkTransport.setAuthorizationHeaderName(config.getAuthorizationHeaderName());
+        for (Map.Entry<String, String> entry : config.getCustomRequestHeaders().entrySet()) {
+            networkTransport.addCustomRequestHeader(entry.getKey(), entry.getValue());
+        }
+    }
+
+    public void setNetworkTransport(OsJavaNetworkTransport transport) {
+        networkTransport = transport;
     }
 
     /**
@@ -36,6 +70,68 @@ public class OsApp implements NativeObject {
                                                                String serviceName) {
         return nativeMakeStreamingRequest(nativePtr, user.getNativePtr(), functionName, arguments, serviceName);
     }
+
+    public OsSyncUser currentUser() {
+        Long userPtr = nativeCurrentUser(nativePtr);
+        return (userPtr != null) ? new OsSyncUser(userPtr) : null;
+    }
+
+    public OsSyncUser[] allUsers() {
+        long[] nativeUsers = nativeGetAllUsers(nativePtr);
+        OsSyncUser[] osSyncUsers = new OsSyncUser[nativeUsers.length];
+
+        for (int i = 0; i < nativeUsers.length; i++) {
+            osSyncUsers[i] = new OsSyncUser(nativeUsers[i]);
+        }
+        return osSyncUsers;
+    }
+
+    public void switchUser(OsSyncUser osUser) {
+        nativeSwitchUser(nativePtr, osUser.getNativePtr());
+    }
+
+    public OsSyncUser login(OsAppCredentials credentials) {
+        AtomicReference<OsSyncUser> success = new AtomicReference<>(null);
+        AtomicReference<AppException> error = new AtomicReference<>(null);
+
+        nativeLogin(nativePtr, credentials.getNativePtr(), new OsJNIResultCallback<OsSyncUser>(success, error) {
+            @Override
+            protected OsSyncUser mapSuccess(Object result) {
+                Long nativePtr = (Long) result;
+                return new OsSyncUser(nativePtr);
+            }
+        });
+
+        return ResultHandler.handleResult(success, error);
+    }
+
+    // Called from JNI
+    @KeepMember
+    public OsJavaNetworkTransport getNetworkTransport() {
+        return networkTransport;
+    }
+
+    private native long nativeCreate(String appId,
+                                     String baseUrl,
+                                     String appName,
+                                     String appVersion,
+                                     long requestTimeoutMs,
+                                     byte[] encryptionKey,
+                                     String syncDirPath,
+                                     String bindingUserInfo,
+                                     String appUserInfo,
+                                     String platform,
+                                     String platformVersion,
+                                     String sdkVersion);
+
+    private static native void nativeLogin(long nativeAppPtr, long nativeCredentialsPtr, OsJavaNetworkTransport.NetworkTransportJNIResultCallback callback);
+
+    @Nullable
+    private static native Long nativeCurrentUser(long nativePtr);
+
+    private static native long[] nativeGetAllUsers(long nativePtr);
+
+    private static native void nativeSwitchUser(long nativeAppPtr, long nativeUserPtr);
 
     private static native long nativeGetFinalizerMethodPtr();
