+++ realm/realm-library/src/syncTestUtils/kotlin/io/realm/TestApp.kt
@@ -36,7 +36,7 @@ class TestApp(
 
     init {
         if (networkTransport != null) {
-            this.networkTransport = networkTransport;
+            this.setNetworkTransport(networkTransport)
         }
     }
